+++ lib/recipes/lifespan-development/recipe.rb
@@ -4,14 +4,27 @@ LIFESPAN_DEVELOPMENT_RECIPE = Kitchen::BookRecipe.new(book_short_name: :lifespan
 do |doc, _resources|
   include Kitchen::Directions
 
+  # Set overrides
+  doc.selectors.override(
+    reference: 'section.references'
+  )
+
   book = doc.book
 
   book.search('cnx-pi').trash
+  metadata = book.metadata
 
   BakePreface.v1(book: book)
 
   BakeUnnumberedFigure.v1(book: book)
   BakeUnnumberedTables.v1(book: book)
+
+  AddInjectedExerciseId.v1(book: book)
+  book.injected_exercises.each do |exercise|
+    BakeInjectedExercise.v1(
+      exercise: exercise
+    )
+  end
 
   BakeChapterTitle.v1(book: book)
   BakeChapterIntroductions.v1(book: book)
@@ -28,6 +41,43 @@ do |doc, _resources|
       BakeNumberedTable.v2(table: table,
                            number: "#{chapter.count_in(:book)}.#{table.count_in(:chapter)}")
     end
+
+    BakeChapterGlossary.v1(chapter: chapter, metadata_source: metadata)
+
+    sections_with_module_links = %w[summary references]
+
+    sections_with_module_links.each do |eoc_section|
+      MoveCustomSectionToEocContainer.v1(
+        chapter: chapter,
+        metadata_source: metadata,
+        container_key: eoc_section,
+        uuid_key: ".#{eoc_section}",
+        section_selector: "section.#{eoc_section}"
+      ) do |section|
+        RemoveSectionTitle.v1(section: section)
+        title = EocSectionTitleLinkSnippet.v1(page: section.ancestor(:page))
+        section.prepend(child: title)
+      end
+    end
+
+    sections_with_exercises = %w[review-questions check-understanding reflection-questions
+                                 media-questions thought-provokers case-study]
+
+    sections_with_exercises.each do |section_key|
+      MoveCustomSectionToEocContainer.v1(
+        chapter: chapter,
+        metadata_source: metadata,
+        container_key: section_key,
+        uuid_key: ".#{section_key}",
+        section_selector: "section.#{section_key}"
+      ) do |section|
+        RemoveSectionTitle.v1(section: section)
+      end
+    end
+
+    BakeAllNumberedExerciseTypes.v1(
+      within: chapter.search('div[data-type="composite-page"]')
+    )
   end
 
   # Appendix
@@ -44,7 +94,9 @@ do |doc, _resources|
     BakeAppendix.v1(page: page, number: appendix_letter)
   end
 
+  BakeIndex.v1(book: book)
   BakeFootnotes.v1(book: book)
+  BakeCompositePages.v1(book: book)
   BakeToc.v1(book: book)
   BakeLinkPlaceholders.v1(book: book)
   BakeFolio.v1(book: book)