+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Fixup `PageElement`, add specs for `CompositePageElement`
+* Bake `index` in `lifespan-development`
+* Bake `eoc` sections in `lifespan-development`
 * Bake `tables` in `lifespan-development`
 * Bake `figures` in `lifespan-development`
 * Create basic files for `lifespan-development` recipe